==> 1.4-WebScraping_adjust_team_composition.ipynb <==
# SCRAPE THE STATUS OF SERIE A PLAYERS FROM TRANSFERMARKT.IT

# import libraries
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import Select
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.common.exceptions import TimeoutException
import unicodedata
import pandas as pd
import numpy as np
import time
import re

# API dataframes
df_players = pd.read_csv('raw_data/all_players_2024_25.csv',
                        index_col=0)
df_teams = pd.read_csv('raw_data/df_serie_a_teams_2024_25.csv',
                      index_col=0)

# DATA PREPARATION

# renaming these two variables (different dataframe, but same variable name)
df_players = df_players.rename(columns={'name': 'player_name'})
df_teams = df_teams.rename(columns={'name': 'team_name'})

# name transformation for the research
def process_player_name(row):
    if '. ' in row['player_name']:
        initial_char = re.split(r'\.\s', row['player_name'])[0]
        last_name = re.split(r'\.\s', row['player_name'])[1]

        if initial_char.lower() == row['first_name'][0].lower():
            first_name = re.split(r'\s', row['first_name'])[0]
            return f'{first_name} {last_name}'

        else:
            return f'{row["first_name"]} {row["last_name"]}'

    return row['player_name']

df_players['processed_player_name'] = df_players.apply(process_player_name, axis=1)

df_players['processed_player_name'].to_list()[:20]

# prefixes removal
def clean_team_name(team_name):
    return re.sub(r'^[A-Z]{1,3}\s', '', team_name)

df_teams['team_name_cleaned'] = df_teams['team_name'].apply(clean_team_name)

serie_A_teams_2024_25 = df_teams['team_name_cleaned'].to_list()
serie_A_teams_2024_25

# join operation
df = pd.merge(df_players[['team_id', 'player_id', 'processed_player_name', 'player_name', 'first_name', 'last_name']],
              df_teams[['team_id', 'team_name', 'team_name_cleaned']],
              how='left',
              on='team_id')

df = df[['player_id', 'processed_player_name', 'player_name', 'first_name', 'last_name',
         'team_id', 'team_name', 'team_name_cleaned']]

df.head()

# special letters convertion into the simple ones
def normalize_special_letters(text):
    special_letter_map = {
                            'æ': 'ae', 'Æ': 'Ae',
                            'œ': 'oe', 'Œ': 'Oe',
                            'ß': 'ss',
                            'Ĳ': 'IJ', 'ĳ': 'ij',
                            'Đ': 'D', 'đ': 'd',
                            'Ł': 'L', 'ł': 'l',
                            'Ø': 'O', 'ø': 'o',
                            'Ŋ': 'N', 'ŋ': 'n'
                        }

    for letter, replacement in special_letter_map.items():
        text = text.replace(letter, replacement)

    # diacritic removal
    normalized = unicodedata.normalize('NFD', text)
    text_without_special_letters = ''.join([char for char in normalized if not unicodedata.combining(char)])

    return text_without_special_letters

# SETTING UP SCRAPING FUNCTIONS

# funtion to close the opened iframe
def iframe_closer():
    try:
        iframe = WebDriverWait(driver, 2).until(
            EC.presence_of_element_located((By.ID, 'sp_message_iframe_953827'))
            )
        driver.switch_to.frame(iframe)
        driver.find_element(By.CLASS_NAME, 'accept-all').click()
        WebDriverWait(driver, 300).until(lambda d: d.execute_script('return document.readyState') == 'complete')
        driver.switch_to.default_content()
    except TimeoutException:
        pass

# funtion to search a player
def search_player(name):
    search_input = WebDriverWait(driver, 2).until(
        EC.element_to_be_clickable((By.CLASS_NAME, 'tm-header__input--search-field'))
    )
    search_input.clear()
    search_input.send_keys(name)
    search_button = WebDriverWait(driver, 2).until(
        EC.element_to_be_clickable((By.CLASS_NAME, 'tm-header__input--search-send'))
    )
    driver.execute_script("arguments[0].click();", search_button)

# funtion that find the way to search a player (with which name format)
def find_player(player, first_name, last_name):
    # full name
    full_name = f'{first_name} {last_name}'

    search_player(full_name)
    WebDriverWait(driver, 300).until(lambda d: d.execute_script('return document.readyState') == 'complete')
    iframe_closer()
    WebDriverWait(driver, 300).until(lambda d: d.execute_script('return document.readyState') == 'complete')

    try:
        team_fields = WebDriverWait(driver, 2).until(
            EC.presence_of_all_elements_located((By.CLASS_NAME, 'tiny_wappen'))
        )
        return
    # standardized name format
    except TimeoutException:
        search_player(player)
        WebDriverWait(driver, 300).until(lambda d: d.execute_script('return document.readyState') == 'complete')
        iframe_closer()
        WebDriverWait(driver, 300).until(lambda d: d.execute_script('return document.readyState') == 'complete')

        try:
            team_fields = WebDriverWait(driver, 2).until(
                EC.presence_of_all_elements_located((By.CLASS_NAME, 'tiny_wappen'))
            )
            return
        # last name
        except TimeoutException:
            search_player(last_name)
            WebDriverWait(driver, 300).until(lambda d: d.execute_script('return document.readyState') == 'complete')
            iframe_closer()
            WebDriverWait(driver, 300).until(lambda d: d.execute_script('return document.readyState') == 'complete')

# function to match the searhed player with the correct one came out from the research
def player_matching_1(team):
    try:
        box = WebDriverWait(driver, 2).until(
                EC.presence_of_all_elements_located((By.CLASS_NAME, 'items'))
            )[0]

        items_status = box.find_elements(By.CLASS_NAME, 'tiny_wappen')
        for item in items_status:
            status = item.get_attribute('title')

            team_lower = team.lower()
            status_lower = status.lower()

            if (team_lower in status_lower and
                  ((team_lower == 'milan' and 'futuro' in status_lower) or
                   (team_lower == 'juventus' and 'next gen' in status_lower) or
                   (team_lower == 'atalanta' and 'u23' in status_lower))):
                return f'{team} second team'

            elif (team_lower in status_lower and
                      ('primavera' in status_lower or
                       re.search(r' u\s*[0-9]+', status_lower) or
                       re.search(r' under\s*[0-9]+', status_lower))):
                return f'{team} youth team'

            elif team_lower in status_lower:
                return team

            elif item == items_status[-1]:
                return 'continue'

    except TimeoutException:
        return 'player not found'

# function to match the searhed player with the correct one came out from the research
def player_matching_2(team):
    serie_A_teams = serie_A_teams_2024_25.copy()
    serie_A_teams.remove(team)
    try:
        box = WebDriverWait(driver, 2).until(
                EC.presence_of_all_elements_located((By.CLASS_NAME, 'items'))
            )[0]

        items_status = box.find_elements(By.CLASS_NAME, 'tiny_wappen')
        for item in items_status:
            status = item.get_attribute('title')
            status_lower = status.lower()

            if any(serie_A_team.lower() in status_lower for serie_A_team in serie_A_teams):

                for serie_A_team in serie_A_teams:
                    if (serie_A_team.lower() in status_lower and
                          ((serie_A_team.lower() == 'milan' and 'futuro' in status_lower) or
                           (serie_A_team.lower() == 'juventus' and 'next gen' in status_lower) or
                           (serie_A_team.lower() == 'atalanta' and 'u23' in status_lower))):
                        return f'{serie_A_team} second team'

                    elif (serie_A_team.lower() in status_lower and
                          ('primavera' in status_lower or
                           re.search(r' u\s*[0-9]+', status_lower) or
                           re.search(r' under\s*[0-9]+', status_lower))):
                        return f'{serie_A_team} youth team'

                    elif serie_A_team.lower() in status_lower:
                        return serie_A_team

            elif item == items_status[-1]:
                return 'continue'

    except TimeoutException:
        return 'player not found'

# function to match the searhed player with the correct one came out from the research
def player_matching_3(team):
    serie_A_teams = serie_A_teams_2024_25.copy()
    serie_A_teams.remove(team)
    try:
        box = WebDriverWait(driver, 2).until(
                EC.presence_of_all_elements_located((By.CLASS_NAME, 'items'))
            )[0]

        items_status = box.find_elements(By.CLASS_NAME, 'tiny_wappen')
        for item in items_status:
            status = item.get_attribute('title')

            team_lower = team.lower()
            status_lower = status.lower()

            if (team_lower in status_lower and
                  ((team_lower == 'milan' and 'futuro' in status_lower) or
                   (team_lower == 'juventus' and 'next gen' in status_lower) or
                   (team_lower == 'atalanta' and 'u23' in status_lower))):
                return f'{team} second team'

            elif (team_lower in status_lower and
                      ('primavera' in status_lower or
                       re.search(r' u\s*[0-9]+', status_lower) or
                       re.search(r' under\s*[0-9]+', status_lower))):
                return f'{team} youth team'

            elif team_lower in status_lower:
                return team

            elif any(serie_A_team.lower() in status_lower for serie_A_team in serie_A_teams):
                for serie_A_team in serie_A_teams:

                    if (serie_A_team.lower() in status_lower and
                          ((serie_A_team.lower() == 'milan' and 'futuro' in status_lower) or
                           (serie_A_team.lower() == 'juventus' and 'next gen' in status_lower) or
                           (serie_A_team.lower() == 'atalanta' and 'u23' in status_lower))):
                        return f'{serie_A_team} second team'

                    elif (serie_A_team.lower() in status_lower and
                          ('primavera' in status_lower or
                           re.search(r' u\s*[0-9]+', status_lower) or
                           re.search(r' under\s*[0-9]+', status_lower))):
                        return f'{serie_A_team} youth team'

                    elif serie_A_team.lower() in status_lower:
                        return serie_A_team

        return 'Team of another league or released or retired'

    except TimeoutException:
        return 'player not found'

# funtion to find the data we need: the player status
def status_research(team):
    try:
        iframe_closer()
        next_page = WebDriverWait(driver, 2).until(
            EC.presence_of_all_elements_located((By.CLASS_NAME, 'tm-pagination__list-item--icon-next-page'))
        )[0]
        while True:
            try:
                next_page = WebDriverWait(driver, 2).until(
                    EC.presence_of_all_elements_located((By.CLASS_NAME, 'tm-pagination__list-item--icon-next-page'))
                )[0]

                final_status = player_matching_1(team)

                if final_status == 'continue':
                    next_page_link = next_page.find_element(By.CLASS_NAME, 'tm-pagination__link')
                    driver.execute_script("arguments[0].click();", next_page_link)
                    WebDriverWait(driver, 300).until(lambda d: d.execute_script('return document.readyState') == 'complete')
                    iframe_closer()
                    WebDriverWait(driver, 300).until(lambda d: d.execute_script('return document.readyState') == 'complete')
                    continue
                else:
                    return final_status

            except TimeoutException:
                final_status = player_matching_1(team)

                if final_status == 'continue':
                    first_page = WebDriverWait(driver, 2).until(
                        EC.presence_of_all_elements_located((By.CLASS_NAME, 'tm-pagination__list-item--icon-first-page'))
                    )[0]
                    driver.execute_script("arguments[0].click();", first_page)
                    WebDriverWait(driver, 300).until(lambda d: d.execute_script('return document.readyState') == 'complete')
                    iframe_closer()
                    WebDriverWait(driver, 300).until(lambda d: d.execute_script('return document.readyState') == 'complete')

                    while True:
                        try:
                            next_page = WebDriverWait(driver, 2).until(
                                EC.presence_of_all_elements_located((By.CLASS_NAME, 'tm-pagination__list-item--icon-next-page'))
                            )[0]

                            final_status = player_matching_2(team)

                            if final_status == 'continue':
                                next_page_link = next_page.find_element(By.CLASS_NAME, 'tm-pagination__link')
                                driver.execute_script("arguments[0].click();", next_page_link)
                                WebDriverWait(driver, 300).until(lambda d: d.execute_script('return document.readyState') == 'complete')
                                iframe_closer()
                                WebDriverWait(driver, 300).until(lambda d: d.execute_script('return document.readyState') == 'complete')
                                continue
                            else:
                                return final_status

                        except TimeoutException:
                            final_status = player_matching_2(team)

                            if final_status == 'continue':
                                return 'player not found'

                            else:
                                return final_status
                else:
                    return final_status

    except TimeoutException:
        return player_matching_3(team)

df['player_status'] = None

subsets = np.array_split(df, 3)

# SCRAPING PHASE

THE PLAYERS LIST WAS DIVIDED INTO THREE PARTS TO REDUCE COMPUTATION TIME AND PREVENT POTENTIAL PERFORMANCE ISSUES

url = 'https://www.transfermarkt.it'
driver = webdriver.Chrome()
driver.implicitly_wait(10)
driver.maximize_window()
driver.get(url)

try:
    time.sleep(3)
    iframe_closer()

    for index, row in subsets[0].iterrows():
        player_name = normalize_special_letters(row['processed_player_name'])
        player_first_name = normalize_special_letters(row['first_name'])
        player_last_name = normalize_special_letters(row['last_name'])

        find_player(player_name, player_first_name, player_last_name)

        df.loc[index, 'player_status'] = status_research(team=row['team_name_cleaned'])

        print(f"{df.loc[index, 'processed_player_name']}: {df.loc[index, 'player_status']}")


finally:
    driver.quit()

url = 'https://www.transfermarkt.it'
driver = webdriver.Chrome()
driver.implicitly_wait(10)
driver.maximize_window()
driver.get(url)

try:
    time.sleep(3)
    iframe_closer()

    for index, row in subsets[1].iterrows():
        player_name = normalize_special_letters(row['processed_player_name'])
        player_first_name = normalize_special_letters(row['first_name'])
        player_last_name = normalize_special_letters(row['last_name'])

        find_player(player_name, player_first_name, player_last_name)

        df.loc[index, 'player_status'] = status_research(team=row['team_name_cleaned'])

        print(f"{df.loc[index, 'processed_player_name']}: {df.loc[index, 'player_status']}")


finally:
    driver.quit()

url = 'https://www.transfermarkt.it'
driver = webdriver.Chrome()
driver.implicitly_wait(10)
driver.maximize_window()
driver.get(url)

try:
    time.sleep(3)
    iframe_closer()

    for index, row in subsets[2].iterrows():
        player_name = normalize_special_letters(row['processed_player_name'])
        player_first_name = normalize_special_letters(row['first_name'])
        player_last_name = normalize_special_letters(row['last_name'])

        find_player(player_name, player_first_name, player_last_name)

        df.loc[index, 'player_status'] = status_research(team=row['team_name_cleaned'])

        print(f"{df.loc[index, 'processed_player_name']}: {df.loc[index, 'player_status']}")


finally:
    driver.quit()

# SOME ANALYSIS AND CORRECTION OF THE ERRORS

df.shape

df.groupby('player_status').count()['processed_player_name']

df.to_csv('C:/Users/dinge/Desktop/UNI/Data Management And Visualization/Data Management/Project/df_all_status_players_2024_25.csv',
          index=False)

df[df['player_status']=='player not found']

Houssem-Eddine Aouar ---> Team of another league or released or retired

Diego Llorente ---> Team of another league or released or retired

Mihajlo Ilić ---> Team of another league or released or retired

Marko Rog ---> Team of another league or released or retired

Christian Dalle Mura ---> Team of another league or released or retired

Seydou Fini ---> Team of another league or released or retired

Eddie Salcedo ---> Team of another league or released or retired

Martín Satriano ---> Team of another league or released or retired

Marley Aké ---> Team of another league or released or retired

Marcos Antônio ---> Team of another league or released or retired

Saná Fernandes ---> Team of another league or released or retired

Natan ---> Team of another league or released or retired

Martin Turk ---> Team of another league or released or retired

Caumenan Ange Emmanu N'Guessan ---> Team of another league or released or retired

Patricio Nehuén Pérez ---> Team of another league or released or retired

Walace ---> Team of another league or released or retired

João Ferreira ---> Team of another league or released or retired

Matheus Martins ---> Team of another league or released or retired

Vivaldo Semedo ---> Team of another league or released or retired

Thomas Henry ---> Team of another league or released or retired
<br>
----------------------------------------------------------
<br>

Christos Papadopoulos ---> Juventus second team

# players not found correction
player_not_found_of_another_league = df[df['player_status']=='player not found']['processed_player_name'].to_list()
player_not_found_of_another_league.remove('Christos Papadopoulos')

for player in player_not_found_of_another_league:
    idx = df[df['processed_player_name']==player].index[0]
    df.loc[idx, 'player_status'] = 'Team of another league or released or retired'

df[df['player_status']=='player not found']

df.loc[407, 'player_status'] = 'Juventus second team'
df[df['processed_player_name']=='Christos Papadopoulos']

# Balotelli is missing
df[df['player_status']=='Genoa']

df_players[df_players['last_name']=='Balotelli']

# one is missing (I don't know who he is)
df[df['player_status']=='Monza']

df_players[df_players['last_name']=='Đurić'][]

# correction
df.loc[636, 'player_status'] = 'Monza'
df[df['player_name']=='M. Đurić']

# there's one extra (I don't know who he is)
df[df['player_status']=='Verona']

# correction
df.loc[254, 'player_status'] = 'Team of another league or released or retired'
df[df['processed_player_name']=='Fabio Rispoli']

df.groupby('player_status').count()['player_id']

df = df['player_id', 'player_status']

df.to_csv('raw_data/df_all_status_players_2024_25.csv',
          index=False)
