==> 2.1-DataStoragePre.ipynb <==
# Import libraries, data and connect do DB

import pandas as pd
from pymongo import MongoClient
import numpy as np

client = MongoClient("mongodb+srv://acorsico:DataMan2023@crypto.mtyohoy.mongodb.net/?retryWrites=true&w=majority")
db = client.SerieA_24_25
teams_collection = db.teams
players_collection = db.players
votes_collection = db.grades

# Load teams' CSV 
teams_df = pd.read_csv('../1 - DataAcquisition/raw_data/serie_a_teams_2024_25.csv')

# Load players' CSV for each team
team_players_dfs = {
    'Lazio': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Lazio_players.csv'),
    'AC Milan': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_AC Milan_players.csv'),
    'Cagliari': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Cagliari_players.csv'),
    'Napoli': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Napoli_players.csv'),
    'Udinese': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Udinese_players.csv'),
    'Genoa': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Genoa_players.csv'),
    'Juventus': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Juventus_players.csv'),
    'AS Roma': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_AS Roma_players.csv'),
    'Atalanta': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Atalanta_players.csv'),
    'Bologna': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Bologna_players.csv'),
    'Fiorentina': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Fiorentina_players.csv'),
    'Torino': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Torino_players.csv'),
    'Verona': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Verona_players.csv'),
    'Inter': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Inter_players.csv'),
    'Empoli': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Empoli_players.csv'),
    'Venezia': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Venezia_players.csv'),
    'Parma': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Parma_players.csv'),
    'Lecce': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Lecce_players.csv'),
    'Como': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Como_players.csv'),
    'Monza': pd.read_csv('../1 - DataAcquisition/raw_data/updated_teams/updated_Monza_players.csv'),
}

# Concat in a unique DF
all_players_df = pd.concat(team_players_dfs.values(), ignore_index = True)
all_players_df.set_index('player_id', inplace = True)
all_players_df

# Load status data
df_status = pd.read_csv('../1 - DataAcquisition/raw_data/df_all_status_players_2024_25.csv')
df_status

teams = ['Milan', 'Roma', 'Atalanta', 'Bologna', 'Genoa',
       'Cagliari', 'Como', 'Empoli', 'Fiorentina', 'Inter',
       'Juventus', 'Lazio', 'Lecce', 'Monza', 'Napoli', 
       'Parma', 'Torino', 'Udinese', 'Venezia', 'Verona']

# FIlter DF to retain only player currently in serie A
df_status = df_status[df_status['player_status'].isin(teams)]
df_status

# Isolate IDs to filter other source of data
keep_ids = df_status.player_id.to_list()
# Filter DF using IDs
all_players_df = all_players_df.loc[all_players_df.index.isin(keep_ids)]
all_players_df

# Save right players list
all_players_df.to_csv('../1 - DataAcquisition/raw_data/all_players_corrected.csv')

## Store teams info and roster

for i, row in teams_df.iterrows():
    team_id = row['team_id']
    team_name = row['name']
    
    # For each team, upload the player data (we assume that the team names are the keys)
    if team_name in team_players_dfs:
        
        players_df = team_players_dfs[team_name]
        players_df = players_df[players_df['player_id'].isin(keep_ids)]
        
        # We create the roster, separated by role
        goalkeepers = players_df[players_df['position'] == 'Goalkeeper'][['player_id', 'name']].to_dict(orient = 'records')
        defenders = players_df[players_df['position'] == 'Defender'][['player_id', 'name']].to_dict(orient = 'records')
        midfielders = players_df[players_df['position'] == 'Midfielder'][['player_id', 'name']].to_dict(orient = 'records')
        attackers = players_df[players_df['position'] == 'Attacker'][['player_id', 'name']].to_dict(orient = 'records')

        # Create doc for each team
        team_document = {
            "_id": team_id,
            "name": team_name,
            "corporate_information": {
                "foundation": row['founded'],
                "city": row['venue_city'],
                "stadium": row['venue_name'],
                "capacity": row['venue_capacity']
            },
            "roster": {
                "goalkeepers": goalkeepers,
                "defenders": defenders,
                "midfielders": midfielders,
                "attackers": attackers
            }
        }
        
        # Insert doc in the MongoDB collection
        teams_collection.insert_one(team_document)

all_players_df

## Store players info

# Iterate on all players
for i, row in all_players_df.iterrows():

    # Define doc structure
    player_document = {

        "_id": i,
        "name": row['name'],
        "position": row['position'],
        "profile": {

            "first_name": row['first_name'],
            "last_name": row['last_name'],
            "birth_date": row['birth_date'],
            "country": row['country'],
            "nationality": row['nationality'],
            "height": row['height'],
            "weight": row['weight']
        }  
    }
    # Add doc to MongoDB collection
    players_collection.insert_one(player_document)

# Checking for errors leading to duplicate indexes
duplicated_indices = all_players_df.index[all_players_df.index.duplicated()]
duplicated_rows = all_players_df.loc[duplicated_indices]
print(duplicated_rows)

## Store fanta stats

# Load fanta data
grades_df = pd.read_csv('../1 - DataAcquisition/raw_data/df_players_grades.csv', index_col = 0)
grades_df

# Creating a dictionary organised by matchdays and teams
documents = {}

for matchday in grades_df['matchday'].unique():
    matchday_data = grades_df[grades_df['matchday'] == matchday]
    
    # Dict for each matchday
    matchday_dict = {}
    
    for team in matchday_data['team_name'].unique():
        team_data = matchday_data[matchday_data['team_name'] == team]
        
        # List of players with their grades and stats
        players = []
        for _, row in team_data.iterrows():
            players.append({
                'Name': row['normalized_player_name'],
                 'Grade': row['grade'],
                 'FantaGrade': row['fantagrade'],
                 'GoalsScored': row['goals_scored'],
                 'GoalsConceded': row['goals_conceded'],
                 'Autogoals': row['autogoals'],
                 'PenaltiesScored': row['penalties_scored'],
                 'PenaltiesMissed': row['penalties_missed'],
                 'PenaltiesSaved': row['penalties_saved'],
                 'Assists': row['assists']  
            })
        
        # Add team with players to the matchday
        matchday_dict[team] = players
    
    # Add matchday to docs
    documents[int(matchday)] = matchday_dict

# Load all docs on MongoDB
for matchday, data in documents.items():
   votes_collection.insert_one({
        '_id': matchday,  # Use matchday as id of the doc
        'teams': data
   })

## Add stats to players collection

# Load stats data
stats = pd.read_csv('../1 - DataAcquisition/raw_data/cleaned_stats.csv', index_col = 0)
stats

# Filter only for correct players
stats = stats[stats.player_id.isin(keep_ids)]
stats

stats.columns

selected_columns = [ 'team_name', 'appereances', 'lineups', 'minutes', 'sub_in', 'sub_out',
       'bench', 'total_shots', 'on_shots', 'scored_goals', 'conceded_goals',
       'saved_goals', 'assists', 'total_passes', 'key_passes',
       'passes_accuracy', 'tackles', 'blocks', 'interceptions', 'total_duels',
       'won_duels', 'attempted_dribbles', 'won_dribbles', 'drawn_fouls',
       'committed_fouls', 'yellow_cards', 'yellowred_cards', 'red_cards',
       'scored_penalties', 'missed_penalties', 'saved_penalties']

#  Group data by player_id including only selected columns
grouped_stats = stats[selected_columns + ["player_id"]].groupby("player_id").apply(
    lambda x: x.to_dict(orient="records")
).to_dict()

# Aggiorna i documenti MongoDB
for player_id, stats in grouped_stats.items():
    players_collection.update_one(
        {"_id": player_id},  # Filtra per _id (player_id corrisponde a _id)
        {"$set": {"stats": stats}},  # Aggiungi il campo "stats"
        upsert=True  # Se il documento non esiste, lo crea
    )

client.close()
