==> 2.2-DataIntegration.ipynb <==
# DATA INTEGRATION (API DATA AND WEBSCRAPING DATA)

# import libraries
import pandas as pd
import re
from rapidfuzz import process, fuzz
import unicodedata

# DATA PREPARATION

<br>

WEB SCRAPING DATA (FANTACALCIO.IT)
<br>
<br>


file_path_scraping = '../1 - DataAcquisition/raw_data/df_players_grades.csv'
df_players_grades_original = pd.read_csv(file_path_scraping)

df_players_grades_original

# order the df by the column 'matchday'
df_players_grades = df_players_grades_original.sort_values(by='matchday')
# perform a groupby on 'player_name' and take the last value of 'team_name' for each group
df_players_grades = df_players_grades.groupby('player_name').agg({'normalized_player_name': 'last','team_name': 'last'}).reset_index()
# add a temporary index
df_players_grades['temporary_player_id'] = df_players_grades.index
# reorder the columns
df_players_grades = df_players_grades[['temporary_player_id', 'player_name', 'normalized_player_name', 'team_name']]

df_players_grades

<br>

API DATA
<br>
<br>

file_path_api_players = '../1 - DataAcquisition/raw_data/all_players.csv'
file_path_api_teams = '../1 - DataAcquisition/raw_data/serie_a_teams_2024_25.csv'

df_api_players = pd.read_csv(file_path_api_players, index_col=0)
df_api_teams = pd.read_csv(file_path_api_teams, index_col=0)

# rename the player name and the team name columns
df_api_players = df_api_players.rename(columns={'name': 'player_name'})
df_api_teams = df_api_teams.rename(columns={'name': 'team_name'})


df_api = pd.merge(df_api_players[['player_id', 'player_name', 'team_id']],
                  df_api_teams[['team_id', 'team_name']],
                  how='left', on='team_id').drop(columns='team_id')
df_api

# team prefixes removal
def clean_team_name(team_name):
    return re.sub(r'^[A-Z]{1,3}\s', '', team_name)

df_api['team_name_cleaned'] = df_api['team_name'].apply(clean_team_name)
df_api['team_name_cleaned'].unique()

<br>

WEB SCRAPING DATA (TRANSFERMARKT.IT)
<br>
<br>

file_path_api_status_players = '../1 - DataAcquisition/raw_data/df_all_status_players_2024_25.csv'
df_api_status_players = pd.read_csv(file_path_api_status_players)

df_api_status_players

# condition needed to take only the correct players
conditions = df_api_status_players['player_status'].isin(['Team of another league or released or retired']) | \
             df_api_status_players['player_status'].str.contains('second team|youth team', case=False, na=False)

player_ids = df_api_status_players.loc[conditions, 'player_id'].tolist()

# filter the API source, keeping only the players who are playing in Serie A this year
df_api = df_api[~df_api['player_id'].isin(player_ids)]
df_api = df_api.drop(columns='team_name')

df_api

<br>

NAMES STANDARDIZATION
<br>
<br>

def add_apostrophe_to_final_accented_vowel(text):
    # set of vowels with grave or acute accents
    accented_vowels = {'à', 'á', 'è', 'é', 'ì', 'í', 'ò', 'ó', 'ù', 'ú'}
    words = text.split()

    # function to add an apostrophe after an accented vowel
    def process_word(word):
        if word[-1] in accented_vowels:
            return f'{word}\''
        return word

    # apply the function to each word
    processed_words = [process_word(word) for word in words]

    return ' '.join(processed_words)

def diacritic_removal(text):
    # NFD decomposition
    normalized = unicodedata.normalize('NFD', text)

    return ''.join([char for char in normalized if not unicodedata.combining(char)])

def normalize_special_letters(text):
    # conversion map of special and combined letters
    special_letter_map = {
        'æ': 'ae', 'Æ': 'Ae',
        'œ': 'oe', 'Œ': 'Oe',
        'ß': 'ss',
        'Ĳ': 'IJ', 'ĳ': 'ij',
        'Đ': 'D', 'đ': 'd',
        'Ł': 'L', 'ł': 'l',
        'Ø': 'O', 'ø': 'o',
        'Ŋ': 'N', 'ŋ': 'n'
    }

    for letter, replacement in special_letter_map.items():
        text = text.replace(letter, replacement)

    # remove any diacritics with NFD normalization
    normalized = unicodedata.normalize('NFD', text)
    text_without_special_letters = ''.join([char for char in normalized if not unicodedata.combining(char)])

    return text_without_special_letters

def name_normalization(row):
    # normalization pipeline
    text = row['player_name']
    text = add_apostrophe_to_final_accented_vowel(text)
    text = diacritic_removal(text)
    text = normalize_special_letters(text)
    return text

df_api['normalized_player_name'] = df_api.apply(name_normalization, axis=1)
df_api = df_api[['player_id', 'player_name', 'normalized_player_name', 'team_name_cleaned']]

df_api

# DATA INTEGRATION

df_players_grades.loc[:, 'player_id_matched'] = None
df_api.loc[:, 'player_id_matched'] = None

common_teams = set(df_api['team_name_cleaned']).intersection(set(df_players_grades['team_name']))
common_teams

<br>

STEP 1
<br>
<br>

matching_results_01 = []

# iterate for each common team
for common_team in common_teams:
    # filter the players of that team in the two DataFrames
    players_df_api = df_api[df_api['team_name_cleaned'] == common_team]
    players_df_grades = df_players_grades[df_players_grades['team_name'] == common_team]

    # for each player in the first DataFrame (players_df_api), look for a match in the second DataFrame (players_df_grades)
    for api_idx, row_api in players_df_api.iterrows():
        player_name_api = row_api['normalized_player_name']

        # find the best match with a threshold of 80
        match = process.extractOne(player_name_api, players_df_grades['normalized_player_name'], scorer=fuzz.ratio, score_cutoff=80)

        # if a match exists, update 'player_id_matched'
        if match:
            matched_name, score, matched_idx = match

            matched_player = players_df_grades.loc[matched_idx]

            # add the result to the list for monitoring
            matching_results_01.append({
                'team_name': common_team,
                'player_id_grades': matched_player['temporary_player_id'],
                'player_name_grades': matched_name,
                'player_name_api': player_name_api,
                'player_id_api': row_api['player_id'],
                'matching_score': score
            })

# DataFrame to visualize the matching results
df_matches = pd.DataFrame(matching_results_01)

df_matches

print(len(df_matches['player_id_grades']), len(df_matches['player_id_grades'].unique()))

print(len(df_matches['player_id_api']), len(df_matches['player_id_api'].unique()))

# update df_players_grades and df_api
for _ , row in df_matches.iterrows():
    id_grades = row['player_id_grades']
    id_api = row['player_id_api']

    df_players_grades.loc[df_players_grades['temporary_player_id'] == id_grades, 'player_id_matched'] = id_api
    df_api.loc[df_api['player_id'] == id_api, 'player_id_matched'] = id_grades

<br>

STEP 2
<br>
<br>

matching_results_02 = []

# iterate for each common team
for common_team in common_teams:
    # filter the players of that team in the two DataFrames (only the players not already matched)
    players_df_api = df_api[(df_api['team_name_cleaned'] == common_team) & (df_api['player_id_matched'].isnull())]
    players_df_grades = df_players_grades[(df_players_grades['team_name'] == common_team) & (df_players_grades['player_id_matched'].isnull())]

    # for each player in the first DataFrame (players_df_api), look for a match in the second DataFrame (players_df_grades)
    for api_idx, row_api in players_df_api.iterrows():
        player_name_api = row_api['normalized_player_name']

        # find the best match with a threshold of 74
        match = process.extractOne(player_name_api, players_df_grades['normalized_player_name'], scorer=fuzz.ratio, score_cutoff=74)

        # if a match exists, update 'player_id_matched'
        if match:
            matched_name, score, matched_idx = match

            matched_player = players_df_grades.loc[matched_idx]

            # add the result to the list for monitoring
            matching_results_02.append({
                'team_name': common_team,
                'player_id_grades': matched_player['temporary_player_id'],
                'player_name_grades': matched_name,
                'player_name_api': player_name_api,
                'player_id_api': row_api['player_id'],
                'matching_score': score
            })

# DataFrame to visualize the matching results
df_matches = pd.DataFrame(matching_results_02)

df_matches

print(len(df_matches['player_id_grades']), len(df_matches['player_id_grades'].unique()))

# update df_players_grades and df_api
for _ , row in df_matches.iterrows():
    id_grades = row['player_id_grades']
    id_api = row['player_id_api']

    df_players_grades.loc[df_players_grades['temporary_player_id'] == id_grades, 'player_id_matched'] = id_api
    df_api.loc[df_api['player_id'] == id_api, 'player_id_matched'] = id_grades

df_api[df_api['normalized_player_name']=='Josep Martinez']

df_api[df_api['normalized_player_name']=='Lautaro Martinez']

df_players_grades[df_players_grades['normalized_player_name']=='L. Martinez']

<br>

STEP 3
<br>
<br>

matching_results_03 = []

# iterate for each common team
for common_team in common_teams:
    # filter the players of that team in the two DataFrames (only the players not already matched)
    players_df_api = df_api[(df_api['team_name_cleaned'] == common_team) & (df_api['player_id_matched'].isnull())]
    players_df_grades = df_players_grades[(df_players_grades['team_name'] == common_team) & (df_players_grades['player_id_matched'].isnull())]

    # for each player in the first DataFrame (players_df_api), look for a match in the second DataFrame (players_df_grades)
    for api_idx, row_api in players_df_api.iterrows():
        player_name_api = row_api['normalized_player_name']

        # find the best match with a threshold of 70
        match = process.extractOne(player_name_api, players_df_grades['normalized_player_name'], scorer=fuzz.ratio, score_cutoff=70)

        # if a match exists, update 'player_id_matched'
        if match:
            matched_name, score, matched_idx = match

            matched_player = players_df_grades.loc[matched_idx]

            # add the result to the list for monitoring
            matching_results_03.append({
                'team_name': common_team,
                'player_id_grades': matched_player['temporary_player_id'],
                'player_name_grades': matched_name,
                'player_name_api': player_name_api,
                'player_id_api': row_api['player_id'],
                'matching_score': score
            })

# DataFrame to visualize the matching results
df_matches = pd.DataFrame(matching_results_03)

df_matches

print(len(df_matches['player_id_api']), len(df_matches['player_id_api'].unique()))

# update df_players_grades and df_api
for _ , row in df_matches.iterrows():
    id_grades = row['player_id_grades']
    id_api = row['player_id_api']

    df_players_grades.loc[df_players_grades['temporary_player_id'] == id_grades, 'player_id_matched'] = id_api
    df_api.loc[df_api['player_id'] == id_api, 'player_id_matched'] = id_grades

<br>

STEP 4
<br>
<br>

matching_results_04 = []

# iterate for each common team
for common_team in common_teams:
    # filter the players of that team in the two DataFrames (only the players not already matched)
    players_df_api = df_api[(df_api['team_name_cleaned'] == common_team) & (df_api['player_id_matched'].isnull())]
    players_df_grades = df_players_grades[(df_players_grades['team_name'] == common_team) & (df_players_grades['player_id_matched'].isnull())]

    # for each player in the first DataFrame (players_df_api), look for a match in the second DataFrame (players_df_grades)
    for api_idx, row_api in players_df_api.iterrows():
        player_name_api = row_api['normalized_player_name']

        # find the best match with a threshold of 50
        match = process.extractOne(player_name_api, players_df_grades['normalized_player_name'], scorer=fuzz.ratio, score_cutoff=50)

        # if a match exists, update 'player_id_matched'
        if match:
            matched_name, score, matched_idx = match

            matched_player = players_df_grades.loc[matched_idx]

            # add the result to the list for monitoring
            matching_results_04.append({
                'team_name': common_team,
                'player_id_grades': matched_player['temporary_player_id'],
                'player_name_grades': matched_name,
                'player_name_api': player_name_api,
                'player_id_api': row_api['player_id'],
                'matching_score': score
            })

# DataFrame to visualize the matching results
df_matches = pd.DataFrame(matching_results_04)

df_matches

OK:

Lazio	219	Gila	Mario Gila	162952

Genoa	359	Messias	Junior Messias	56396

Juventus	570	Yildiz	K. Yıldız	339883

Monza	374	Mota	Dany Mota	30603

Monza	338	Mari'	Pablo Mari'	46792

Cagliari	23	Azzi	Paulo Azzi	56207

Roma	380	  N'Dicka 	E. Ndicka	1807

filtered_df_matches = df_matches[df_matches['player_id_grades'].isin([219, 359, 570, 374, 338, 23, 380])]

print(len(filtered_df_matches['player_id_grades']), len(filtered_df_matches['player_id_grades'].unique()))

# update df_players_grades and df_api
for _ , row in filtered_df_matches.iterrows():
    id_grades = row['player_id_grades']
    id_api = row['player_id_api']

    df_players_grades.loc[df_players_grades['temporary_player_id'] == id_grades, 'player_id_matched'] = id_api
    df_api.loc[df_api['player_id'] == id_api, 'player_id_matched'] = id_grades

<br>

STEP 5
<br>
<br>

matching_results_05 = []


# filter the players who have not yet been matched
players_df_api = df_api[df_api['player_id_matched'].isnull()]
players_df_grades = df_players_grades[df_players_grades['player_id_matched'].isnull()]

# for each player in the first DataFrame (players_df_api), look for a match in the second DataFrame (players_df_grades)
for api_idx, row_api in players_df_api.iterrows():
    player_name_api = row_api['normalized_player_name']

    # find the best match in the first DataFrame with a threshold of 80
    match = process.extractOne(player_name_api, players_df_grades['normalized_player_name'], scorer=fuzz.ratio, score_cutoff=80)

    # if a match exists, update 'player_id_matched'
    if match:
        matched_name, score, matched_idx = match

        matched_player = players_df_grades.loc[matched_idx]

        # add the result to the list for monitoring
        matching_results_05.append({
            'team_name_grades': matched_player['team_name'],
            'player_id_grades': matched_player['temporary_player_id'],
            'player_name_grades': matched_name,
            'player_name_api': player_name_api,
            'player_id_api': row_api['player_id'],
            'team_name_api': row_api['team_name_cleaned'],
            'matching_score': score
        })

# DataFrame to visualize the matching results
df_matches = pd.DataFrame(matching_results_05)

df_matches


NO:

Atalanta	175	E. Toure'	I. Toure'	195893	Udinese

Juventus	299	L. Cerri	A. Cerri	30570	Como

Sassuolo	45	Berardi	A. Berardi	30912	Verona

filtered_df_matches = df_matches[~df_matches['player_id_grades'].isin([175, 299, 45])]

print(len(filtered_df_matches['player_id_grades']), len(filtered_df_matches['player_id_grades'].unique()))

duplicates = filtered_df_matches['player_id_grades'][filtered_df_matches['player_id_grades'].duplicated()]

duplicate_rows = filtered_df_matches[filtered_df_matches['player_id_grades'].isin(duplicates)]

duplicate_rows

filtered_df_matches = filtered_df_matches[~filtered_df_matches['player_id_api'].isin([307103, 128338])]

print(len(filtered_df_matches['player_id_grades']), len(filtered_df_matches['player_id_grades'].unique()))

print(len(filtered_df_matches['player_id_api']), len(filtered_df_matches['player_id_api'].unique()))

# update df_players_grades and df_api
for _ , row in filtered_df_matches.iterrows():
    id_grades = row['player_id_grades']
    id_api = row['player_id_api']

    df_players_grades.loc[df_players_grades['temporary_player_id'] == id_grades, 'player_id_matched'] = id_api
    df_api.loc[df_api['player_id'] == id_api, 'player_id_matched'] = id_grades

<br>

STEP 6
<br>
<br>

matching_results_06 = []


# filter the players who have not yet been matched
players_df_api = df_api[(df_api['player_id_matched'].isnull())]
players_df_grades = df_players_grades[df_players_grades['player_id_matched'].isnull()]

# for each player in the first DataFrame (players_df_api), look for a match in the second DataFrame (players_df_grades)
for api_idx, row_api in players_df_api.iterrows():
    player_name_api = row_api['normalized_player_name']

    # find the best match in the first DataFrame with a threshold of 72
    match = process.extractOne(player_name_api, players_df_grades['normalized_player_name'], scorer=fuzz.ratio, score_cutoff=72)

    # if a match exists, update 'player_id_matched'
    if match:
        matched_name, score, matched_idx = match

        matched_player = players_df_grades.loc[matched_idx]

        # add the result to the list for monitoring
        matching_results_06.append({
            'team_name_grades': matched_player['team_name'],
            'player_id_grades': matched_player['temporary_player_id'],
            'player_name_grades': matched_name,
            'player_name_api': player_name_api,
            'player_id_api': row_api['player_id'],
            'team_name_api': row_api['team_name_cleaned'],
            'matching_score': score
        })
# DataFrame to visualize the matching results
df_matches = pd.DataFrame(matching_results_06)
df_matches = df_matches[df_matches['matching_score']<80]

df_matches

NO:

Sassuolo	365	K. Miranda	Juan Miranda	134	Bologna

Empoli	496	S. Shpendi	S. Sensi	30535	Monza

Milan	87	Camarda	D. Camara	91424	Parma

filtered_df_matches = df_matches[~df_matches['player_id_grades'].isin([365, 496, 87])]

print(len(filtered_df_matches['player_id_grades']), len(filtered_df_matches['player_id_grades'].unique()))

# update df_players_grades and df_api
for _ , row in filtered_df_matches.iterrows():
    id_grades = row['player_id_grades']
    id_api = row['player_id_api']

    df_players_grades.loc[df_players_grades['temporary_player_id'] == id_grades, 'player_id_matched'] = id_api
    df_api.loc[df_api['player_id'] == id_api, 'player_id_matched'] = id_grades

<br>

STEP 7
<br>
<br>

matching_results_07 = []


# filter the players who have not yet been matched
players_df_api = df_api[(df_api['player_id_matched'].isnull()) & (~df_api['player_id'].isin([134, 30535, 91424]))]
players_df_grades = df_players_grades[df_players_grades['player_id_matched'].isnull()]

# for each player in the first DataFrame (players_df_api), look for a match in the second DataFrame (players_df_grades)
for api_idx, row_api in players_df_api.iterrows():
    player_name_api = row_api['normalized_player_name']

    # find the best match in the first DataFrame with a threshold of 60
    match = process.extractOne(player_name_api, players_df_grades['normalized_player_name'], scorer=fuzz.ratio, score_cutoff=60)

    # if a match exists, update 'player_id_matched'
    if match:
        matched_name, score, matched_idx = match

        matched_player = players_df_grades.loc[matched_idx]

        # add the result to the list for monitoring
        matching_results_07.append({
            'team_name_grades': matched_player['team_name'],
            'player_id_grades': matched_player['temporary_player_id'],
            'player_name_grades': matched_name,
            'player_name_api': player_name_api,
            'player_id_api': row_api['player_id'],
            'team_name_api': row_api['team_name_cleaned'],
            'matching_score': score
        })

# DataFrame to visualize the matching results
df_matches = pd.DataFrame(matching_results_07)
df_matches = df_matches[df_matches['matching_score'] <= 72]

OK:

Lecce	515	Strefezza	Gabriel Strefezza	31624	Como

Genoa	343	Martin	Aaron Martin	25914	Genoa

Salernitana	162	Dia	B. Dia	22015.0	Lazio

Fiorentina	18	Arthur Melo	  Arthur	150	Juventus

Juventus	387	Nicolussi Caviglia	H. Nicolussi	881	Venezia

filtered_df_matches = df_matches[(df_matches['player_id_grades'].isin([515, 343, 162, 18, 387]))]

print(len(filtered_df_matches['player_id_grades']), len(filtered_df_matches['player_id_grades'].unique()))

filtered_df_matches

filtered_df_matches = filtered_df_matches[~filtered_df_matches['player_id_api'].isin([30663, 237066])]

print(len(filtered_df_matches['player_id_grades']), len(filtered_df_matches['player_id_grades'].unique()))

print(len(filtered_df_matches['player_id_api']), len(filtered_df_matches['player_id_api'].unique()))

# update df_players_grades and df_api
for _ , row in filtered_df_matches.iterrows():
    id_grades = row['player_id_grades']
    id_api = row['player_id_api']

    df_players_grades.loc[df_players_grades['temporary_player_id'] == id_grades, 'player_id_matched'] = id_api
    df_api.loc[df_api['player_id'] == id_api, 'player_id_matched'] = id_grades

<br>

STEP 8
<br>
<br>

matching_results_08 = []


# filter the players who have not yet been matched
players_df_api = df_api[(df_api['player_id_matched'].isnull()) & (~df_api['player_id'].isin([134, 30535, 91424]))]
players_df_grades = df_players_grades[df_players_grades['player_id_matched'].isnull()]

# for each player in the first DataFrame (players_df_api), look for a match in the second DataFrame (players_df_grades)
for api_idx, row_api in players_df_api.iterrows():
    player_name_api = row_api['normalized_player_name']

    # find the best match in the first DataFrame with a threshold of 60
    match = process.extractOne(player_name_api, players_df_grades['normalized_player_name'], scorer=fuzz.ratio, score_cutoff=50)

    # if a match exists, update 'player_id_matched'
    if match:
        matched_name, score, matched_idx = match

        matched_player = players_df_grades.loc[matched_idx]

        # add the result to the list for monitoring
        matching_results_08.append({
            'team_name_grades': matched_player['team_name'],
            'player_id_grades': matched_player['temporary_player_id'],
            'player_name_grades': matched_name,
            'player_name_api': player_name_api,
            'player_id_api': row_api['player_id'],
            'team_name_api': row_api['team_name_cleaned'],
            'matching_score': score
        })

# DataFrame to visualize the matching results
df_matches = pd.DataFrame(matching_results_08)
df_matches = df_matches[df_matches['matching_score'] <= 60]

<br>

NOT MATCHED PLAYERS
<br>
<br>

len(df_players_grades[df_players_grades['player_id_matched'].isnull()])

len(df_api[df_api['player_id_matched'].isnull()])

# NAME HARMONIZATION

df_players_grades_original['api_player_id'] = None

# filter the df (only matched players)
df_players_grades_not_nulls = df_players_grades[~df_players_grades['player_id_matched'].isnull()]

for idx, row in df_players_grades_not_nulls.iterrows():
    # player name and team name (from df_players_grades source) extraction
    player_name = row['player_name']
    team_name = row['team_name']

    # player name (from df_api source) extraction
    player_id_matched = row['player_id_matched']
    condition = df_api['player_id']==player_id_matched
    api_player_name = df_api.loc[condition, 'player_name'].iloc[0]

    # substitution of the name in df_players_grades_original
    conditions = (df_players_grades_original['player_name'] == player_name)

    df_players_grades_original.loc[conditions, 'player_name'] = api_player_name
    df_players_grades_original.loc[conditions, 'api_player_id'] = player_id_matched

df_players_grades_original = df_players_grades_original[['api_player_id', 'player_name',
                                                         'normalized_player_name', 'team_name', 'role',
                                                         'grade', 'fantagrade', 'goals_scored',
                                                         'goals_conceded', 'autogoals', 'penalties_scored',
                                                         'penalties_missed', 'penalties_saved', 'assists',
                                                         'player_of_the_match', 'season', 'matchday']]

df_players_grades_original

len(set(df_players_grades_original['api_player_id']) - set([None]))

df_players_grades_original.to_csv('../1 - DataAcquisition/raw_data/df_players_grades_post_integration.csv',
                                  index=False)

df_players_grades_original
