==> 3.3-Query.ipynb <==
# Import libraries
import pandas as pd
from pymongo import MongoClient
import numpy as np

# Connect to MongoDB database
client = MongoClient("mongodb+srv://acorsico:DataMan2023@crypto.mtyohoy.mongodb.net/?retryWrites=true&w=majority")
db = client.SerieA_24_25
players_collection = db.players_post
grades_collection = db.grades_post

players = players_collection.find({}, {"name": 1, "stats.scored_goals": 1})

results = []

for player in players:
    name = player.get("name", "Unknown")
    stats = player.get("stats", [])
    
    total_goals = sum(stat.get("scored_goals", 0) for stat in stats)
    
    results.append({"name": name, "total_goals": total_goals})

ordered_results = sorted(results, key=lambda x: x["total_goals"], reverse=True)

print("Players with most goals")
for result in ordered_results:
    print(f"{result['name']}: {result['total_goals']} goals")


media_fantagrade = {}

documents = grades_collection.find({})

for document in documents:
    teams = document.get('teams', {})  
    for team_name, players in teams.items():  
        for player in players:  
            name = player.get("Name", "Sconosciuto") 
            
            if "FantaGrade" in player:
                if name not in media_fantagrade:
                    media_fantagrade[name] = []
                media_fantagrade[name].append(player["FantaGrade"])

# Calcolare la media per ciascun giocatore
media_fantagrade_finale = {
    name: sum(grades) / len(grades) for name, grades in media_fantagrade.items() if grades
}

# Ordinare i risultati in ordine decrescente per media FantaGrade
media_fantagrade_ordinati = sorted(media_fantagrade_finale.items(), key=lambda x: x[1], reverse=True)

# Stampare i risultati ordinati
print("")
for player, average in media_fantagrade_ordinati:
    print(f"{player}: {average:.2f}")

client.close()
